split up some constraint types
