+++ ark/type/__tests__/defaults.test.ts
@@ -1,11 +1,12 @@
 import { attest, contextualize } from "@ark/attest"
 import { scope, type } from "arktype"
-import type { Date, Default } from "arktype/internal/keywords/ast.ts"
+import type { Default } from "arktype/internal/keywords/ast.ts"
 import { invalidDefaultKeyKindMessage } from "arktype/internal/parser/objectLiteral.ts"
 import {
-	shallowDefaultMessage,
-	writeNonLiteralDefaultMessage
+    shallowDefaultMessage,
+    writeNonLiteralDefaultMessage
 } from "arktype/internal/parser/string/shift/operator/default.ts"
+import type { Date } from "../keywords/constructors/Date.ts"
 
 contextualize(() => {
 	describe("parsing and traversal", () => {
