+++ ark/type/__tests__/get.test.ts
@@ -1,11 +1,8 @@
 import { attest, contextualize } from "@ark/attest"
 import { writeInvalidKeysMessage, writeNumberIndexMessage } from "@ark/schema"
 import { ark, type } from "arktype"
-import type {
-	Matching,
-	constrain,
-	string
-} from "arktype/internal/keywords/ast.ts"
+import type { constrain, string } from "arktype/internal/keywords/ast.ts"
+import type { Matching } from "arktype/internal/keywords/string/string.ts"
 
 contextualize(() => {
 	it("can get shallow roots by path", () => {
