+++ ark/type/__tests__/pipe.test.ts
@@ -8,12 +8,8 @@ import {
 	type ArkErrors
 } from "@ark/schema"
 import { ark, scope, type, type Type } from "arktype"
-import type {
-	MoreThan,
-	Out,
-	To,
-	constrain
-} from "arktype/internal/keywords/ast.ts"
+import type { Out, To, constrain } from "arktype/internal/keywords/ast.ts"
+import type { MoreThan } from "arktype/internal/keywords/number/number.ts"
 
 contextualize(() => {
 	it("base", () => {
