+++ ark/type/keywords/number/epoch.ts
@@ -1,6 +1,5 @@
 import { rootNode } from "@ark/schema"
-import type { AtLeast, AtMost, DivisibleBy } from "../ast.ts"
-import type { number } from "./number.ts"
+import type { AtLeast, AtMost, DivisibleBy, number } from "./number.ts"
 
 /**
  * As per the ECMA-262 specification:
