+++ package.json
@@ -16,7 +16,7 @@
 	"private": true,
 	"scripts": {
 		"prChecks": "pnpm lint && pnpm build && pnpm testRepo && pnpm bench && pnpm testTsVersions",
-		"attest": "pnpm tsx ./ark/attest/cli/cli.ts",
+		"attest": "ts ./ark/attest/cli/cli.ts",
 		"build": "pnpm -r build",
 		"buildCjs": "ARKTYPE_CJS=1 pnpm -r build",
 		"rmBuild": "pnpm -r exec rm -rf out",
